AI Tweaks
Starting world generation after this, will be using a .tgf format to
store graph information, and a separate .txt file to store information
about each node. Spawn points will be generated upon visitation of an
outpost. Thus when leaving said outpost, enemies will be spawned when
traveling to the next outpost.

You can have one outpost tracked at a time, and you will be able to
iterate through all discovered outposts that are adjacent to the nearest
visited outpost. Green is discovered but not visited, blue is discovered
and not visited.